working on selected display, serial_num and quantities.

+++ pages/model_list.php
@@ -3,12 +3,20 @@
 require_once '../includes/class_autoloader.inc.php';
 
 $modelCtrl = new ModelCtrl();
-$models = $modelCtrl->getAllModels(); // Fetch all models
+
+if($user_role != 'admin' && $storeId != '') {
+$models = $modelCtrl->getModelsInStoreWithQuantity($storeId);
+} else {
+$models = $modelCtrl->getAllModelsWithQuantity(); // Fetch all models
+}
+
 ?>
 
     <div class="view-store-container">
     <div class="create_container">
+        <?php if ($user_role === 'admin') : ?>
     <a href="dashboard.php?page=create_model" class="create-link">Create New Model</a>
+            <?php endif; ?>
 </div>
         <h2 class="h2">Models List</h2>
         <table class="store-table">
@@ -33,7 +41,7 @@ $models = $modelCtrl->getAllModels(); // Fetch all models
                             <td><?php echo htmlspecialchars($model['model_name']); ?></td>
                             <td><?php echo htmlspecialchars($model['brand']); ?></td>
                             <td><?php echo htmlspecialchars($model['category']); ?></td>
-                            <td><?php echo htmlspecialchars($model['quantity']); ?></td>
+                            <td><?php echo $storeId . ': ' .$modelCtrl->getQuantityByStore($model['model_id'], $storeId) ?></td>
                             <td><?php echo htmlspecialchars($model['specification']); ?></td>
                             <td>
                                 <?php
@@ -46,7 +54,12 @@ $models = $modelCtrl->getAllModels(); // Fetch all models
                             <td>
                                 <!-- Edit button -->
                                 <a class="action-link" href="dashboard.php?page=edit_model&model_id=<?php echo $model['model_id']; ?>">Edit</a>
-                                <a class="action-link delete" href="../includes/delete_model.inc.php?id=<?php echo $model['model_id']; ?>">Delete</a>
+                                <?php if ($user_role === 'admin') : ?>
+                                    <a class="action-link delete" href="../includes/delete_model.inc.php?id=<?php echo $model['model_id']; ?>" onclick="return confirm('Are you sure to Delete this Model?')">Delete Model</a>
+                                <?php else : ?>
+                                    <a class="action-link delete" href="dashboard.php?page=remove_equipments&model_id=<?php echo $model['model_id']; ?>">Delete Equipments</a>
+                                <?php endif; ?>
+                                <a class="action-link More" href="dashboard.php?page=equipment_list&model_id=<?php echo $model['model_id']; ?>">More</a>
                             </td>
                         </tr>
                     <?php endforeach; ?>
